data_engineering_workshop/requirements.txt
kafka-python
psycopg2-binary
SQLAlchemy
pandas
boto3
pyarrow
great_expectations
dbt-core
dbt-postgres